2024-07-15 feat: Code Refactoring

+++ src/main/java/com/backend/vocanote/service/UserService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -22,21 +23,33 @@ public class UserService {
         return userRepository.findById(id).orElse(null);
     }
 
-    public void createUser(UserDTO userDTO) {
+    public boolean createUser(UserDTO userDTO) {
         User user = new User();
         user.setName(userDTO.getName());
         user.setEmail(userDTO.getEmail());
         userRepository.save(user);
+        return true;
     }
 
-    public void updateUser(Long id, UserDTO userDTO) {
-        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
-        user.setName(userDTO.getName());
-        user.setEmail(userDTO.getEmail());
-        userRepository.save(user);
+    public boolean updateUser(Long id, UserDTO userDTO) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            User existingUser = user.get();
+            existingUser.setName(userDTO.getName());
+            existingUser.setEmail(userDTO.getEmail());
+
+            userRepository.save(existingUser);
+            return true;
+        }
+        return false;
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
